rename classify subjects prompt
