+++ tools/text/prompts/__init__.py
@@ -1,6 +1,6 @@
 # flake8: noqa
 from .classify_video import *
-from .classify_subject import *
+from .classify_subjects import *
 from .lecture_summary import *
 from .parse_image import *
 from .queries import *
